Eliminate some msvc warnings

+++ qceplib-qwt/qwt-6.1.3/src/qwt_series_data.cpp
@@ -78,7 +78,7 @@ QRectF qwtBoundingRectT(
         from = 0;
 
     if ( to < 0 )
-        to = series.size() - 1;
+        to = (int)(series.size() - 1);
 
     if ( to < from )
         return boundingRect;
